refactor: Move Unitful from direct dependency to weak dependency with new extension
- Rename DimensionalDataExt to HAPIClientDimensionalDataExt for clarity

chore: migrate from CompatHelper to Dependabot for Julia dependencies
