+++ Project.toml
@@ -3,6 +3,9 @@ uuid = "0ce0b37c-8851-4c5b-98a3-e691a5f19c53"
 version = "0.2.2"
 authors = ["Zijin Zhang <zzj956959688@gmail.com> and contributors"]
 
+[workspace]
+projects = ["docs", "test"]
+
 [deps]
 CSV = "336ed68f-0bac-5ca0-87d4-7b16caf5d00b"
 Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
@@ -10,13 +13,14 @@ HTTP = "cd3eb016-35fb-5094-929b-558a96fad6f3"
 JSON = "682c06a0-de6a-54ab-a142-c8b1cf79cde6"
 SpaceDataModel = "0b37b92c-f0c5-4a52-bd5c-390dec20857c"
 Tables = "bd369af6-aec1-5ad0-b16a-f7cc5008161c"
-Unitful = "1986cc42-f94f-5a68-af5c-568840ba703d"
 
 [weakdeps]
 DimensionalData = "0703355e-b756-11e9-17c0-8b28908087d0"
+Unitful = "1986cc42-f94f-5a68-af5c-568840ba703d"
 
 [extensions]
-DimensionalDataExt = "DimensionalData"
+HAPIClientDimensionalDataExt = "DimensionalData"
+HAPIClientUnitfulExt = "Unitful"
 
 [compat]
 CSV = "0.10"
