+++ test/Project.toml
@@ -5,3 +5,4 @@ HAPIClient = "0ce0b37c-8851-4c5b-98a3-e691a5f19c53"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 TestItemRunner = "f8b46487-2199-4994-9208-9a1283c18c0a"
 TestItems = "1c621080-faea-4a02-84b6-bbd5e436b8fe"
+Unitful = "1986cc42-f94f-5a68-af5c-568840ba703d"
